# contracts/careerasana_openapi_v1.yaml
openapi: 3.1.0
info:
  title: CareerĀsanā API
  version: 1.7.0
  summary: Workflow-driven API for the CareerĀsanā Career Operating System.
  description: |
    The CareerĀsanā API exposes **business capabilities**, not CRUD tables. It is the contract
    over Blueprint v1.3.1 and Database Design Spec v1.3.

    ## Design tenets
    - **Workflow-driven.** Endpoints name capabilities (`/activation/complete`, `/blueprint/regenerate`,
      `/readiness/recalculate`, `/outcomes/{id}/confirm`) rather than persistence (`/createGoal`).
    - **The Twin is never mutated directly.** It changes only via `POST /twin/signals`; the Coach
      emits signals, it does not write the Twin. There is intentionally no `PUT /twin`.
    - **Provenance & confidence are first-class.** Every Twin fact carries `confidence {value, basis}`
      and provenance (`source` → signal, `capturedAt`). Readiness carries a `specVersion`; AI artifacts
      carry generation metadata.
    - **Asynchrony is explicit.** LLM-bound capabilities (`/blueprint/regenerate`) return `202` with a
      job; cheap pure computations (`/readiness/recalculate`) return `200` synchronously.

    ## Security
    Bearer JWT issued by Supabase Auth. All endpoints require auth unless noted. Tenant isolation is
    enforced server-side by Postgres **Row-Level Security** keyed to the authenticated user; the API
    never exposes another user's rows regardless of parameters.

    ## Errors
...
    `errors[]` array for field-level validation.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/session:
    post:
      tags: [Authentication]
      operationId: createSession
      summary: Establish or refresh a session
      description: |
        Exchanges credentials or a refresh token for an access/refresh token pair via Supabase Auth.
        Most clients use the Supabase SDK directly; this endpoint documents the contract for first-party clients.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionRequest' }
            examples:
              password:
                value: { grantType: password, email: asha@example.com, password: "••••••••" }
              refresh:
                value: { grantType: refresh_token, refreshToken: "rt_abc123" }
      responses:
        '200':
          description: Session established.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase-issued access token. Row-Level Security scopes every query to the token's user.
...
  headers:
...
    RateLimit-Limit:
      description: Request quota for the current window (IETF draft). Present on all responses.
      schema: { type: integer }
    RateLimit-Remaining:
      description: Requests remaining in the current window.
      schema: { type: integer }
    RateLimit-Reset:
      description: Seconds until the rate-limit window resets.
      schema: { type: integer }
    Retry-After:
      description: Seconds to wait before retrying.
      schema: { type: integer }
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { type: "https://errors.careerasana.com/bad-request", title: Bad Request, status: 400, detail: Malformed JSON body. }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { type: "https://errors.careerasana.com/unauthorized", title: Unauthorized, status: 401, detail: Bearer token missing or expired. }
...
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After: { $ref: '#/components/headers/Retry-After' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimit-Limit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimit-Remaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimit-Reset' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { type: "https://errors.careerasana.com/rate-limit", title: Too Many Requests, status: 429 }
    ServerError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { type: "https://errors.careerasana.com/internal", title: Internal Server Error, status: 500 }
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 Problem Details, extended with operational fields.
      properties:
        type: { type: string, format: uri, description: A URI identifying the problem type. }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string, description: Stable machine-readable error code. }
        requestId: { type: string, description: Matches the X-Request-Id response header for support correlation. }
        correlationId: { type: string, description: Trace id correlating related requests. }
        timestamp: { type: string, format: date-time }
        documentation: { type: string, format: uri, description: Link to docs for this error type. }
      required: [type, title, status]
...
    SessionRequest:
      type: object
      properties:
        grantType: { type: string, enum: [password, refresh_token] }
        email: { type: string, format: email }
        password: { type: string }
        refreshToken: { type: string }
      required: [grantType]
    Session:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, enum: [bearer] }
        expiresIn: { type: integer, description: Seconds until access token expiry. }
      required: [accessToken, tokenType, expiresIn]
...